move behaviour interfaces to right folder
