+++ Runtime/BehavioursInterfaces/IInitialSpawn.cs
@@ -1,4 +1,4 @@
-namespace ReupVirtualTwin.behaviours
+namespace ReupVirtualTwin.behaviourInterfaces
 {
     public interface IInitialSpawn
     {
