+++ Runtime/BehavioursInterfaces/IMaintainHeight.cs
@@ -1,4 +1,4 @@
-namespace ReupVirtualTwin.behaviours
+namespace ReupVirtualTwin.behaviourInterfaces
 {
     public interface IMaintainHeight
     {
